requirements.txt
azure-core
azure-search-documents
azure-identity
python-dotenv
requests
fastapi
uvicorn
openai
pillow
PyPDF2
pytesseract
python-docx
python-multipart